clutter-gester-action: Cancel gesture when a device gets grabbed
If an actor takes a grab on an input device, that's a good indicator
that the user expects input to be handled solely by the actor and
any gestures that are taking place aren't meant for the stage.

At the moment, the gesture code has no insight into when a grab
gets taken, though. If a grab does get taken it missed out on the
motion and button release events that have now been redirected to
the actor.

Those missed events can confuse the event processing code into thinking
the user is doing a multi-press gesture, when the user is actually just
e.g., doing a series of clicks.

This commit addresses the problem by installing an event filter that
can see the events before they're delivered to the grabbed actor.

The filter checkes if a device a grabbed, and if so cancels any
in-flight gestures that are currently getting processed.