clutter/actor: Notify first/last child changes when moving children
Both removing and adding a child may change a container's first/last
child property, however operations that combine both currently only
emit change notifications for the latter.

This is particularly problematic for gnome-shell, because StWidget's
relies on first/last-child notifications to compute the first / last
visible widgets and to apply on them the relative pseudo classes that
can be used by CSS selectors.
And so, in the described case we might end up having a container with
more than one widget, where a single widget is marked as both first and
last visible one, breaking theming.

Thus, use REMOVE_CHILD_NOTIFY_FIRST_LAST flag to ensure that we notify
changes to first/last child, thawing the object first to make sure that
we only emit the notifications once the actors reordering is completed
and not while the container is missing an actor.

Add tests simulating the case, and further checks on first/last-child
notifications.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2047>