mdk/pipewire: Add way to temporarily only process PipeWire
This will be used to in a blocking fashion wait for new frames from the
PipeWire producer end. It's done by recreating the GSource and attaching
it to a temporary GMainContext. This allows "roundtrip" like
functionality to be added to users of MdkPipewire.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1949>