wayland: Check serial for activation if no surface is provided
Having a "requester" surface helps with accuracy (e.g. ensuring the
pointer is still within coordinates, besides the token serial still
remaining valid). In the case we don't have one, still check the
serial so that we know no other button presses/etc happened in between
and activate the window if none happened.