udev-rules: Only disable modifiers on i915
The only currently known issue with allocating scanout buffers with
modifiers is memory constraints in multi head setups. Heuristics for
handling that are not implemented, but since it doesn't apply to
anything but i915, remove the other drivers from the deny list.

The other drivers had modifiers disabled to marginally increase the
chance of becoming scanout-able when allocated by Xwayland.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2146>